third_party/blink/web_tests/virtual/dialog-light-dismiss-disabled/external/wpt/html/semantics/interactive-elements/the-dialog-element/dialog-popover-closedby-complex.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] clicking outside all with modeless dialogA and modeless dialogB
  assert_equals: Second dialog should be closed expected false but got true
[FAIL] clicking popoverA with modeless dialogA and modeless dialogB
  assert_equals: Second dialog should be closed expected false but got true
[FAIL] clicking dialogA with modeless dialogA and modeless dialogB
  assert_equals: Second dialog should be closed expected false but got true
[FAIL] clicking outside all with modeless dialogA and modal dialogB
  assert_equals: Second dialog should be closed expected false but got true
[FAIL] clicking popoverA with modeless dialogA and modal dialogB
  assert_equals: Second dialog should be closed expected false but got true
[FAIL] clicking dialogA with modeless dialogA and modal dialogB
  assert_equals: Second dialog should be closed expected false but got true
[FAIL] clicking outside all with modal dialogA and modeless dialogB
  assert_equals: Second dialog should be closed expected false but got true
[FAIL] clicking popoverA with modal dialogA and modeless dialogB
  assert_equals: Second dialog should be closed expected false but got true
[FAIL] clicking dialogA with modal dialogA and modeless dialogB
  assert_equals: Second dialog should be closed expected false but got true
[FAIL] clicking outside all with modal dialogA and modal dialogB
  assert_equals: Second dialog should be closed expected false but got true
[FAIL] clicking popoverA with modal dialogA and modal dialogB
  assert_equals: Second dialog should be closed expected false but got true
[FAIL] clicking dialogA with modal dialogA and modal dialogB
  assert_equals: Second dialog should be closed expected false but got true
Harness: the test ran to completion.
